rename tool to tools
