+++ streetview/__init__.py
@@ -1,3 +1,4 @@
 from .api import get_panorama_meta, get_streetview  # noqa
 from .download import get_panorama, get_panorama_async  # noqa
 from .search import search_panoramas  # noqa
+from .tools import crop_bottom_and_right_black_border  # noqa
